III-5851 - Show RecentLocations next to Pick a location
