+++ src/i18n/de.json
@@ -203,7 +203,8 @@
       "recent_locations": {
         "title": "Wählen Sie einen Standort aus, den Sie kürzlich verwendet haben",
         "info": "Wir haben die Standorte, für die Sie kürzlich eingegeben haben, hier für Sie gespeichert. Sie können sie mit einem Klick hinzufügen.",
-        "other": "Oder wählen Sie einen anderen Standort"
+        "other": "Oder wählen Sie einen anderen Standort",
+        "pick": "Wählen Sie einen Standort aus"
       },
       "validation_messages": {
         "street_and_number": {
