+++ src/i18n/fr.json
@@ -203,7 +203,8 @@
       "recent_locations": {
         "title": "Choisissez un lieu que vous avez utilisé récemment",
         "info": "Nous avons mis les lieux pour lesquels vous avez récemment entré ici pour vous. Vous pouvez les ajouter en un clic.",
-        "other": "Ou choisissez un autre lieu"
+        "other": "Ou choisissez un autre lieu",
+        "pick": "Choisissez un lieu"
       },
       "validation_messages": {
         "street_and_number": {
